==> notebooks/2 - Case Study.ipynb <==
<hr>
# PYTHON FOR DATA SCIENCE
### Exploring a Real World Example

#Agenda

- <p style="color: red">Define the problem and the approach</p>
- Data basics: loading data, looking at your data, basic commands
- Handling missing values
- Intro to scikit-learn
- Grouping and aggregating data
- Feature selection
- Fitting and evaluating a model
- Deploying your work


##In this notebook you will

- Determine if the problem is worth solving
- Define an approach
- Develop a workflow outline

<hr>
## Revisiting Kaggle's Give Me Some Credit Competition

### Intro

#### We'll be exploring an age-old prediction problem--predicting risk on consumer loans.

#### We've selected this topic as a case study because the problem is well-defined and familiar to most people. Additionally, this is a binary classification problem, so the strategies we apply should be relatively extensible to other problems you may encounter totally unrelated to credit and finance.

### About the data

The data is made available to us by Kaggle and was used in a competition in 2011.

[http://www.kaggle.com/c/GiveMeSomeCredit](http://www.kaggle.com/c/GiveMeSomeCredit)

### About the prediction problem

Predict the probability that somebody will experience financial distress in the next two years.

## Developing an understanding of the data and the problem

#### Key questions we'll need to keep in mind

- How do losses occur?
- What are the characteristics that constitute credit default?
- How often do people "go bad?"
- How might we improve loss rates?

import pandas as pd
df = pd.read_csv("./data/credit-training.csv")
df.head()

df.SeriousDlqin2yrs.mean()

##Building a Model
Since we have a fairly large volume of data (150K), we're going to build a predictive model that will enable us to automatically score each applicant. We will provide each applicant with a credit score. This will give us an easy to interpret, human readable form of the model.

##Our Strategy
If we're building a model, we're going to need a way to know whether or not it's working. Convincing other is oftentimes the most challenging parts of an analysis. Making repeatable, well documented work with clear success metrics makes all the difference.

For our classifier, we're going to use the following build methodology:

from IPython.core.display import Image
Image(url="https://s3.amazonaws.com/demo-datasets/traintest.png")

##Other Things to consider

- [Precision and Recall](http://en.wikipedia.org/wiki/Precision_and_recall) - What good is our classifier if declines everyone?
- [Overfitting](http://en.wikipedia.org/wiki/Overfitting) - Is your model describing noise or signal?
- [Algorithms](http://en.wikipedia.org/wiki/Statistical_classification#Algorithms) - What type of classifiers might work in this scenario?

##We just did the following

- Determined we had a problem in which there was ample room for improvement
- Decided we could use a predictive model to decrease losses
- Proposed a high level workflow
