==> notebooks/5 - Imputing Data.ipynb <==
#Agenda

- Define the problem and the approach
- Data basics: loading data, looking at your data, basic commands
- <p style="color: red">Handling missing values</p>
- Intro to scikit-learn
- Grouping and aggregating data
- Feature selection
- Fitting and evaluating a model
- Deploying your work

#In this workbook you will
- Learn about imputation and what it's used for
- Use the KNearestNeighbors algorithm to impute data
- Come up with your own imputation strategy

#Imputation
Imputation is fairly intuitive. For the missing data in our dataset, we're going to replace it with values that come from similar records in our dataset that aren't null.
To do this we're going to use the NearestNeighbors algorithm.

import pandas as pd
import numpy as np
import pylab as pl

df = pd.read_csv("./data/credit-data-post-import.csv")

##Cross Validation
We're going to use the simplest type of cross validation. we'll simply split our data into 2 groups: training and test. we'll use the training set to calibrate our model and then use the test set to  evaluate how effective it is.

is_test = np.random.uniform(0, 1, len(df)) > 0.75
train = df[is_test==False]
test = df[is_test==True]

len(train), len(test)

##Be sure to calibrate the imputation with the training set

from sklearn.neighbors import KNeighborsRegressor

income_imputer = KNeighborsRegressor(n_neighbors=1)

#split our data into 2 groups; data containing nulls and data 
# not containing nulls we'll train on the latter and make
# 'predictions' on the null data to impute monthly_income
train_w_monthly_income = train[train.monthly_income.isnull()==False]
train_w_null_monthly_income = train[train.monthly_income.isnull()==True]

train_w_monthly_income.corr()

train_w_monthly_income.corr().ix[:,5]

cols = ['number_real_estate_loans_or_lines', 'number_of_open_credit_lines_and_loans']
income_imputer.fit(train_w_monthly_income[cols], train_w_monthly_income.monthly_income)

##Replace the mising values

new_values = income_imputer.predict(train_w_null_monthly_income[cols])


train_w_null_monthly_income['monthly_income'] = new_values
new_values

#combine the data back together
train = train_w_monthly_income.append(train_w_null_monthly_income)
len(train)

test['monthly_income_imputed'] = income_imputer.predict(test[cols])
test.head()

test['monthly_income'] = np.where(test.monthly_income.isnull(), test.monthly_income_imputed,
                                  test.monthly_income)

print pd.value_counts(train.monthly_income.isnull())
print pd.value_counts(test.monthly_income.isnull())

train.to_csv("./data/credit-data-trainingset.csv", index=False)
test.to_csv("./data/credit-data-testset.csv", index=False)

##Trying your own imputation strategy
Partner with the person next to you and think of other ways you might be able to impute missing valus for the `monthly_income` column.

Things to consider:

- Are `number_of_open_credit_lines_and_loans` and `number_real_estate_loans_or_lines` the best predictors?
- What are some other methods you could use for replacing nulls?
- What about handling outlying values? Should a montly income of $3MM be treated as missing?
- Check out the [scikit-learn docs](http://scikit-learn.org/stable/auto_examples/imputation.html) for more examples
- How might you evaluate the effectiveness of your imputation?

##We just did the following

- Split our data into a training set for building our model and a test set for evaluating its performance
- Used KNearestNeighbors to fill in missing values for `monthly_income`
