==> notebooks/6 - Aggregation and Grouping.ipynb <==
#Agenda

- Define the problem and the approach
- Data basics: loading data, looking at your data, basic commands
- Handling missing values
- Intro to scikit-learn
- <p style="color: red">Grouping and aggregating data</p>
- Feature selection
- Fitting and evaluating a model
- Deploying your work

#In this workbook you will
- Receive and overview of the `apply` function
- Write custom functions for analyzing data in `pandas`
- Do SQL style joins on your data frames

import pandas as pd
import numpy as np
import pylab as pl

df = pd.read_csv("./data/credit-data-trainingset.csv")
df.head()

##Apply
"Applies" or operates on a column in your data frame with a given function. This is analagous to an Excel formula.

df.monthly_income.apply(np.log)

###Applying with lambda functions
A `lambda` function is an anonymous function. Think of it just as a shorthand way to define a quick function that you need once.

add_10 = lambda x: x + 10
plus = lambda x, y: x + y

print add_10(9)
print plus(10, 20)

df.monthly_income.apply(lambda x: np.log(x + 1))

#numpy actually has log(x + 1)
help(np.log1p)

###Using custom functions
If you can't do it in a one-liner lambda function don't worry. `pandas` also let's `apply` your own custom functions. You can use custom functions when applying on Series and also when operating on chunks of data frames in `groupby`s.

def inverse(x):
    return 1 / (x + 1)

df.monthly_income.apply(inverse)

####Write a custom function called `cap_value(x, cap)` that will set x to the cap if x > cap. Then apply it to debt_ratio with a cap of 5.

def cap_value(x, cap):
    """
    x - a value
    cap - threshold value for x; if x > cap, then x is set to cap
    Examples:
        cap_value(1000, 10)
        10
        cap_value(10, 100)
        10
        """
    # your code here
    return None
print cap_value(1000, 10)==10
print cap_value(10, 100)==10
print df.debt_ratio.apply(lambda x: cap_value(x, 5.0)).mean()#should be close to 1.28

##Split -> Apply ->Combine
Split, Apply, Combine is a data munging methodology similar in spirit to `SQL`'s `GROUP BY`. The idea being you split your data into chunks, operate on those chunks, and then combine the results together into a single table. `groupby` in `pandas` works exactly the same way. But since we're using Python and not SQL, we have a lot more flexibility in terms of the types of operations we can perform in the apply step.

From the `pandas` documentation:

 - Splitting the data into groups based on some criteria
 - Applying a function to each group independently
 - Combining the results into a data structure

###Split

subset = df[['serious_dlqin2yrs', 'age', 'monthly_income']]
subset.groupby("serious_dlqin2yrs")

###Apply / Combine
Aggregate whatever is returned

subset.groupby("serious_dlqin2yrs").mean()

What's really going on here? You can see below that when you `groupby` a certain variable(s), you're literally splitting the data into chunks based on each possible value of that variable.

for name, group in subset.groupby("serious_dlqin2yrs"):
    print "splitting by: ", name
    print group.mean()
    print "*"*80

####Use groupby to calculate the percent of customers that went bad for each age

###You can also aggregate by multiple functions

subset.groupby("serious_dlqin2yrs").agg([np.min, np.mean, np.median, np.max])

###`pandas` also let's you use custom apply functions

def age_x_income(frame):
    x = (frame.age * frame.monthly_income)
    return np.mean(x)

subset.groupby("serious_dlqin2yrs").apply(age_x_income)

###Merging and Joining

pop = pd.read_csv("./data/uspop.csv")
pop

cols = ['age', 'monthly_income', 'serious_dlqin2yrs']
result = pd.merge(df[cols] , pop, how='left', on='age')
result

len(result) > len(df)

pd.value_counts(pop.age).head()

pop = pop[pop.age.duplicated()==False]

cols = ['age', 'monthly_income', 'serious_dlqin2yrs']
joined = pd.merge(df[cols] , pop, how='left', on='age')

pop.tail()

joined.est_pop = joined.est_pop.fillna(4957.0)
joined.est_pop.describe()

##`pandasql`
Training wheels for `pandas`. We developed and open sourced `pandasql` to help people coming from other languages ease into the `pandas` syntax. It allows you to query `pandas` data frames like they were `SQL` tables.

from pandasql import sqldf
pysqldf = lambda q: sqldf(q, globals())

query = """
select
    serious_dlqin2yrs
    , sum(1) as total
from
    df
group by
    serious_dlqin2yrs;
"""
pysqldf(query)

query = """
select
    age
    , avg(serious_dlqin2yrs) as pct_delinquent
from
    df
group by
    age
order by
    age;
"""
pysqldf(query)

##We just did the following

- Used `apply` to make custom data transformations
- Did `groupby` and aggregate operations using `pandas`
- Used both `pandas` and `pandasql` to merge data frames together
