==> notebooks/9 - Deployment.ipynb <==
#Agenda

- Define the problem and the approach
- Data basics: loading data, looking at your data, basic commands
- Handling missing values
- Intro to scikit-learn
- Grouping and aggregating data
- Feature selection
- Fitting and evaluating a model
- <p style="color: red">Deploying your work</p>


#In this workbook you will
- Deploy your model to Yhat
- Autogenerate a webapp with your model's data"

##Use the same code we wrote earlier

import pandas as pd
import numpy as np
import pylab as pl

train = pd.read_csv("./data/credit-data-trainingset.csv")
test = pd.read_csv("./data/credit-data-testset.csv")

from sklearn.tree import DecisionTreeClassifier
features = ['revolving_utilization_of_unsecured_lines', 'debt_ratio',
            'monthly_income', 'age', 'number_of_times90_days_late']
clf = DecisionTreeClassifier(min_samples_leaf=1000)
clf.fit(train[features], train.serious_dlqin2yrs)

##You can use any custom function you want

def convert_prob_to_score(p):
    """
    takes a probability and converts it to a score
    Example:
        convert_prob_to_score(0.1)
        466
    """
    odds = (1 - p) / p
    scores = np.log(odds)*(40/np.log(2)) + 340
    return scores.astype(np.int)

##Deploying to Yhat
from yhat import BaseModel, Yhat


yh = Yhat("YOUR USERNAME", "YOUR APIKEY")

class LoanModel(BaseModel):
    def transform(self, newdata):
        df = pd.DataFrame(newdata)
        return df

    def predict(self, df):
        data = df[self.features]
        result = {}
        p = self.clf.predict_proba(data)
        p = p[::, 1]
        score = convert_prob_to_score(p)
        result["prob"] = p
        result["score"] = score
        return result


loan_model = LoanModel(clf=clf, features=features,
                 udfs=[convert_prob_to_score])

##Executing the model
####Raw Data > `transform` > `predict` > Back to Client

testcase = test.head(1).to_dict('list')
transform_result = loan_model.transform(testcase)
loan_model.predict(transform_result)

yh.deploy("DataGothamLoanModel", loan_model)

##Create a test case with one of the holdout samples and run it through your api. See if the results match

testcase = test.head(100).tail(1)
yh.predict("DataGothamLoanModel", 4, testcase.to_dict('list'))

prob = clf.predict_proba(testcase[features])
convert_prob_to_score(prob[::,1])

##Generate a single page app with your model/data

yh.document("DataGothamLoanModel", 4, train[features])

##We just did the following

- Deployed our model to Yhat
- Switched models on the fly
- Generated a webapp for our model using `yhat.document`
